+++ client/src/modules/auth/index.js
@@ -1,9 +1,12 @@
 import LoginContainer from './login/LoginContainer';
 import SignupContainer from './signup/SignupContainer';
 import authReducer from './reducers';
+import { AuthViews, NonAuthViews } from './wrapper';
 
 export {
   authReducer,
+  AuthViews,
   LoginContainer,
+  NonAuthViews,
   SignupContainer
 };
