+++ client/src/modules/auth/wrapper/index.js
@@ -1,0 +1,4 @@
+import AuthViews from './AuthViews';
+import NonAuthViews from './NonAuthViews';
+
+export { AuthViews, NonAuthViews };
